requirements.txt
pandas==0.25.3
discord.py==1.5.1
discord==1.0.1
PyYAML==5.3.1
